Continue factoring StripeRequest
Move more request properties into `StripeRequest`
