+++ stripe/src/test/java/com/stripe/android/StripeRequestMatcher.java
@@ -3,11 +3,14 @@ package com.stripe.android;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 
+import com.stripe.android.exception.InvalidRequestException;
+
 import org.mockito.ArgumentMatcher;
 
+import java.io.UnsupportedEncodingException;
 import java.util.Map;
 
-public class StripeRequestMatcher implements ArgumentMatcher<StripeRequest> {
+class StripeRequestMatcher implements ArgumentMatcher<StripeRequest> {
     @NonNull private final StripeRequest.Method mMethod;
     @NonNull private final String mUrl;
     @Nullable private final Map<String, ?> mParams;
@@ -36,8 +39,14 @@ public class StripeRequestMatcher implements ArgumentMatcher<StripeRequest> {
 
     @Override
     public boolean matches(@NonNull StripeRequest request) {
+        final String url;
+        try {
+            url = request.getUrl();
+        } catch (UnsupportedEncodingException | InvalidRequestException e) {
+            return false;
+        }
         return mMethod == request.method &&
-                mUrl.equals(request.url) &&
+                mUrl.equals(url) &&
                 (mParams == null || mParams.equals(request.params)) &&
                 (mOptions == null || mOptions.equals(request.options));
     }
